replace direct core lib usage with SciChartReact in LiDAR3DPointCloudDemo component
